fix(admin): cast AVG result to float8 in metrics query
PostgreSQL's AVG(EXTRACT(EPOCH FROM ...)) returns NUMERIC, but sqlx
expects f64 (FLOAT8). Add explicit ::float8 cast to fix 500 errors
on the /metrics endpoint.

+++ packages/admin/src/metrics.rs
@@ -62,7 +62,7 @@ pub async fn fetch_metrics(pool: &PgPool) -> Result<MetricsSnapshot, sqlx::Error
     .await?;
 
     let avg_duration: (Option<f64>,) = sqlx::query_as(
-        "SELECT AVG(EXTRACT(EPOCH FROM (completed_at - started_at))) \
+        "SELECT AVG(EXTRACT(EPOCH FROM (completed_at - started_at)))::float8 \
          FROM jobs WHERE status = 'completed' \
          AND completed_at > NOW() - INTERVAL '24 hours'",
     )
